Move CTacticsConfiguration traits
